fix wp for working whit multi site and can save api urls

+++ classes/WP_MTPC.php
@@ -12,11 +12,12 @@ class WP_MTPC extends stdClass {
 	public function __construct($plugin_dir) {
 		$options = get_option('metapic_options');
 		$this->debugMode = (defined("MTPC_DEBUG") && MTPC_DEBUG === true);
+        $this->debugMode= true; //overwrite line above
 		$this->plugin_dir = $plugin_dir;
 		$this->plugin_url = plugins_url() . '/' . basename(__DIR__);
-		$apiUrl = ($this->debugMode && isset($options["uri_string"]) && $options["uri_string"] != "") ? $options["uri_string"] : "http://metapic-testapi.herokuapp.com/";
-		$this->client = new ApiClient($apiUrl, get_site_option("mtpc_api_key"), get_site_option("mtpc_secret_key"));
-		$this->setupOptionsPage();
+        $this->client = new ApiClient( get_site_option("mtpc_api_url"), get_site_option("mtpc_api_key"), get_site_option("mtpc_secret_key"));
+       // echo json_encode($this->client->getUsers());
+        $this->setupOptionsPage();
 		$this->setupLang();
 		$this->setupNetworkOptions();
 		$this->setupIframeRoutes();
@@ -40,16 +41,16 @@ class WP_MTPC extends stdClass {
 	}
 
 	private function setupOptionsPage() {
-		add_action('admin_init', function () {
-			$options = get_option('metapic_options');
+      	add_action('admin_init', function () {
+            $options = get_option('metapic_options');
 			register_setting('metapic_options', 'metapic_options', function ($input) {
-				$options = get_option('metapic_options');
+			    $options = get_option('metapic_options');
 				switch ($_POST["mtpc_action"]) {
 					case "reactivate":
 						$adminEmail = get_bloginfo("admin_email");
 						$user = $this->client->activateUser($adminEmail);
 						if (isset($user["id"])) {
-							update_option("mtpc_active_account", true);
+            				update_option("mtpc_active_account", true);
 							update_option("mtpc_access_token", $user["access_token"]["access_token"]);
 						}
 						break;
@@ -61,8 +62,7 @@ class WP_MTPC extends stdClass {
 			});
 
 			$activeAccount = get_option("mtpc_active_account");
-
-			if (!$activeAccount) {
+            if (!$activeAccount) {
 				add_settings_section('plugin_main', 'Login', function () {
 					echo '<p>Please login to your Metapic account</p>';
 				}, 'plugin');
@@ -105,26 +105,39 @@ class WP_MTPC extends stdClass {
 	}
 
 	private function updateOptions($options, $input) {
-		$options['uri_string'] = trim($input['uri_string'], "/");
+     	$options['uri_string'] = trim($input['uri_string'], "/");
+
 		if (!get_option("mtpc_active_account")) {
-			$options['email_string'] = trim($input['email_string']);
-			$options['password_string'] = trim($input['password_string']);
-			try {
-				$user = $this->client->login($options['email_string'], $options['password_string']);
-				if ($user) {
-					update_option("mtpc_active_account", true);
-					update_option("mtpc_access_token", $user["access_token"]["access_token"]);
-					$this->setStatusMessage("Login successful");
-				}
-				else {
-					throw new Exception;
-				}
-			}
-			catch (Exception $e) {
-				delete_option("mtpc_active_account");
-				delete_option("mtpc_access_token");
-				$this->setStatusMessage("Invalid username or password", "error");
-			}
+
+            if (! is_multisite() ) {
+                $options['email_string'] = trim($input['email_string']);
+                $options['password_string'] = trim($input['password_string']);
+                try {
+                    $user = $this->client->login($options['email_string'], $options['password_string']);
+                    if ($user) {
+                        update_option("mtpc_active_account", true);
+                        update_option("mtpc_access_token", $user["access_token"]["access_token"]);
+                        $this->setStatusMessage("Login successful");
+                    } else {
+                        throw new Exception;
+                    }
+                } catch (Exception $e) {
+                    delete_option("mtpc_active_account");
+                    delete_option("mtpc_access_token");
+                    $this->setStatusMessage("Invalid username or password", "error");
+                }
+            }else{
+                $adminEmail = get_bloginfo("admin_email");
+                $user = $this->client->activateUser($adminEmail);
+
+                if($user["access_token"]==null){
+                    $this->client->createUser(array("email"=>$adminEmail));
+                    $user = $this->client->activateUser($adminEmail);
+                }
+                update_option("mtpc_active_account", true);
+                update_option("mtpc_access_token",$user["access_token"]["access_token"] );
+                $this->setStatusMessage("Login successful");
+            }
 		}
 		return $options;
 	}
@@ -139,14 +152,21 @@ class WP_MTPC extends stdClass {
 		add_action( 'network_admin_menu', function() {
 			add_submenu_page('settings.php', 'Metapic', 'Metapic', 'manage_network', 'metapic', function() {
 				if ( $_POST ) {
-					$_GET['updated'] = true;
-					update_site_option("mtpc_api_key", $_POST["api_key"]);
+                  	$_GET['updated'] = true;
+    				update_site_option("mtpc_api_key", $_POST["api_key"]);
 					update_site_option("mtpc_secret_key", $_POST["secret_key"]);
-					$this->client = new ApiClient("http://api.metapic.dev", $_POST["api_key"], $_POST["secret_key"]);
+                    //echo json_encode($_POST);
+                    if(isset($_POST["API_url"])) {
+                        $apiUrl=$_POST["API_url"];
+                    }else{
+                        $apiUrl="http://metapic-testapi.herokuapp.com";
+                    }
+                    update_site_option("mtpc_api_url",$apiUrl);
+                    $this->client = new ApiClient($apiUrl, $_POST["api_key"], $_POST["secret_key"]);
 					$isValid = $this->client->checkClient($_POST["api_key"]);
 					update_site_option("mtpc_valid_client", ($isValid["status"] == 200));
 				}
-				$this->getTemplate("metapic-site-options");
+				$this->getTemplate("metapic-site-options",array("debugMode"=>$this->debugMode));
 			});
 		});
 	}
@@ -181,10 +201,11 @@ class WP_MTPC extends stdClass {
 
 		add_action('parse_request', function ($wp) {
 			if (array_key_exists('metapic_randomNummber', $wp->query_vars)) {
-				wp_send_json([
-					"access_token" => ["access_token" => get_option("mtpc_access_token")],
-					"metapicApi" => $this->client->getBaseUrl()
-				]);
+
+               wp_send_json([
+                    "access_token" => ["access_token" => get_option("mtpc_access_token")],
+                    "metapicApi" => $this->client->getBaseUrl()
+                ]);
 			}
 			return;
 		});
